+++ src/common/parser.ts
@@ -1012,6 +1012,18 @@ export class Parser extends EventEmitter {
                     case 'inverse':
                       tmp.style |= FontStyle.Inverse;
                       break;
+                    case 'hidden':
+                      tmp.style |= FontStyle.Hidden;
+                      break;
+                    case 'strikeout':
+                      tmp.style |= FontStyle.Strikeout;
+                      break;
+                    case 'overline':
+                      tmp.style |= FontStyle.Overline;
+                      break;
+                    case 'doubleunderline':
+                      tmp.style |= FontStyle.DoubleUnderline;
+                      break;
                   }
                 }
                 break;